requirements/requirements.txt
rich
pydantic
pyclamd
pillow